processing/pyproject.toml
[project]
name = "processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.7",
    "kafka-python>=2.0.2",
    "uvicorn>=0.34.0",
]
